fix: Merge light and dark shadow in one drawable
